Standalone executable built for Bash3k Pre-Alpha 1 Minor changes to allow freezing of files
git-svn-id: svn://svn.code.sf.net/p/oblivionworks/code/Programs/Wrye%20Bash@1859 c80116c5-1073-4541-b2a9-63b9fd661f84

+++ experimental/Bash3k/bash/games/SKY/manager.py
@@ -246,7 +246,8 @@ class GameManager:
             out.write('# This file is used to tell Skyrim which data files to load.\n\n')
             for modName in mods_list:
                 out.write('%s\n' % (modName,))
-        os.remove(backup_name)
+        if os.path.exists(backup_name):
+            os.remove(backup_name)
         os.rename(self.active_files_path, backup_name)
         os.rename(temp_name, self.active_files_path)
